Moved modal to dashboard rather than app catalog
